Move organization Impls that depend on diesel feature into diesel mod
This will limit the number if `#[cfg(feature = "diesel")]` statements needed to fully
encapsulate the diesel feature.
